+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add the possibility to delete a coupon from the backoffice.
 - module list is now reversed. Delivery modules appear first, then payment and finally classic modules.
 - display a loader when a module is uploaded
+- Change product prices export and import format to be compatible, now using product_sale_elements id as key to identify PSE.
 
 # 2.1.1